+++ src/antidote/_internal/utils/__init__.py
@@ -16,15 +16,6 @@ class Default(enum.Enum):
 
 
 @API.private
-class YesSet(Set[object]):
-    def __init__(self) -> None:
-        super().__init__()
-
-    def __contains__(self, item: object) -> bool:
-        return True
-
-
-@API.private
 class Copy(enum.Enum):
     IDENTICAL = enum.auto()
 
@@ -37,4 +28,4 @@ class Copy(enum.Enum):
 
 
 __all__ = ['debug_repr', 'short_id', 'FinalImmutable', 'Immutable', 'AbstractMeta',
-           'FinalMeta', 'SlotRecord', 'API', 'Default', 'Copy', 'YesSet']
+           'FinalMeta', 'SlotRecord', 'API', 'Default', 'Copy']